[WIP] MatrixHandler
