+++ src/models/IrcAction.ts
@@ -26,7 +26,7 @@ type IrcActionType = "message"|"emote"|"topic"|"notice";
 export class IrcAction {
     constructor (
         public readonly type: IrcActionType,
-        public readonly text: string,
+        public text: string,
         public readonly ts: number = 0 ) {
         if (ACTION_TYPES.indexOf(type) === -1) {
             throw new Error("Unknown IrcAction type: " + type);
@@ -39,14 +39,16 @@ export class IrcAction {
             case "emote":
             case "notice":
                 if (matrixAction.htmlText) {
-                    let ircText = ircFormatting.htmlToIrc(matrixAction.htmlText);
+                    const text = ircFormatting.htmlToIrc(matrixAction.htmlText);
+                    const ircText = text !== null ? text : matrixAction.text; // fallback if needed.
                     if (ircText === null) {
-                        ircText = matrixAction.text; // fallback
+                        throw Error("ircText is null");
                     }
                     // irc formatted text is the main text part
                     return new IrcAction(matrixAction.type, ircText, matrixAction.ts)
+                } else if (matrixAction.text) {
+                    return new IrcAction(matrixAction.type, matrixAction.text, matrixAction.ts);
                 }
-                return new IrcAction(matrixAction.type, matrixAction.text, matrixAction.ts);
             case "image":
                 return new IrcAction(
                     "emote", "uploaded an image: " + matrixAction.text, matrixAction.ts
@@ -58,7 +60,9 @@ export class IrcAction {
             case "file":
                 return new IrcAction("emote", "posted a file: " + matrixAction.text, matrixAction.ts);
             case "topic":
-                return new IrcAction(matrixAction.type, matrixAction.text, matrixAction.ts);
+                if (matrixAction.text) {
+                    return new IrcAction(matrixAction.type, matrixAction.text, matrixAction.ts);
+                }
             default:
                 log.error("IrcAction.fromMatrixAction: Unknown action: %s", matrixAction.type);
                 return null;
